+++ src/lib.rs
@@ -5,7 +5,7 @@ pub use context::HeifContext;
 pub use encoder::*;
 pub use enums::*;
 pub use errors::{HeifError, HeifErrorCode, HeifErrorSubCode, Result};
-pub use image::Image;
+pub use image::*;
 pub use image_handle::{ImageHandle, ItemId};
 pub use reader::{Reader, StreamReader};
 pub use utils::check_file_type;